Fix circular import in ui
