+++ src/halfpipe/ui/step.py
@@ -5,11 +5,30 @@
 from abc import abstractmethod
 from collections import defaultdict
 from copy import deepcopy
+from pathlib import Path
 from typing import ClassVar, Dict, Optional, Type
 
+from ..ingest.database import Database
 from ..logging import logger
-from .base import Context
+from ..model.spec import Spec, SpecSchema
 from .components import SingleChoiceInputView, SpacerView, TextElement, TextView
+
+
+class Context:
+    def __init__(self) -> None:
+        spec_schema = SpecSchema()
+        spec = spec_schema.load(spec_schema.dump({}), partial=True)
+        assert isinstance(spec, Spec)
+        self.spec: Spec = spec  # initialize with defaults
+        self.database = Database(self.spec)
+        self.workdir: Path | None = None
+        self.use_existing_spec = False
+        self.debug = False
+        self.already_checked: set[str] = set()
+
+    def put(self, fileobj):
+        self.database.put(fileobj)
+        return len(self.spec.files) - 1
 
 
 class Step:
